AlgVex-server-backup-2026-04-14/web/backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Authentication
authlib>=1.3.0
httpx>=0.26.0
itsdangerous>=2.1.2
python-jose[cryptography]>=3.3.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Binance API
python-binance>=1.0.19

# Utilities
pydantic>=2.5.3
pydantic-settings>=2.1.0

# CORS
starlette>=0.35.0

# Performance Metrics (v3.0.2 - use official libraries, Python 3.12 compatible)
# See: docs/research/EVALUATION_FRAMEWORK.md
# CRITICAL FIX: Use empyrical-reloaded instead of empyrical (Python 3.12 compatible)
# - empyrical (original) is unmaintained since 2019 and breaks with Python 3.12
# - empyrical-reloaded is actively maintained and does NOT force pandas-datareader dependency
# - pandas-datareader is INCOMPATIBLE with Python 3.12 - DO NOT install it
# Reference: https://github.com/stefan-jansen/empyrical-reloaded
empyrical-reloaded>=0.5.12,<1.0  # Sharpe/Sortino/Calmar/MDD/VaR (Python 3.12 compatible)
scipy>=1.9.0,<2.0                # Bootstrap, statistical tests
statsmodels>=0.14.0,<1.0         # Multiple hypothesis testing correction
pandas>=2.2.2,<4.0               # Time series processing (required for numpy>=2.0)
numpy>=1.24.0,<3.0               # Numerical computing
